feat(routes): Add routing configuration for App
A new routing configuration was set up for the application in app-routes.tsx. This was done to introduce a separation of anonymous/public routes and protected/private routes. The 'AuthProvider' component is used to manage authentication status. Private routes are accessible with "/dashboard/*" path and all other paths are treated as public routes.