feat: Remove unused code and simplify modal closure logic
This commit removes the `reset` property in the `combobox.tsx` that was not being used in favor of cleaner code. Similarly, removed the use of `useEffect` and `randomId()` function in `sideover.tsx` as they were unnecessary.

In `useOpenAndCloseModalBasedRoute.ts`, the code has been refactored to simplify the closure logic of the modal. The updated logic does not require a `navigateBack` parameter and makes the code easier to read and maintain.